# abacusflow-portal/abacusflow-portal-web/src/main/resources/static/openapi.yaml
openapi: 3.0.0
info:
  title: abacusflow api
  version: 0.0.1
  description: 小算盘服务接口文档
servers:
  - url: http://localhost:8080
...
security:
  - bearerAuth: []
paths:
...
  /me/invitations/accept:
    post:
      operationId: acceptTenantInvitation
      tags:
        - tenant
      summary: Accept a tenant invitation with the current verified identity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AcceptTenantInvitationInput"
      responses:
        "200":
          description: Invitation accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantInvitation"
        "400":
          description: 邀请无效、邮箱不匹配或邮箱未验证
  /tenant:
    get:
      operationId: getCurrentTenant
      tags:
        - tenant
      summary: Get selected tenant detail
      responses:
        "200":
          description: Tenant detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantDetail"
        "403":
          description: 无权限
        "404":
          description: 租户不存在

    put:
      operationId: updateCurrentTenant
      tags:
        - tenant
      summary: Update selected tenant profile
      description: 需要 tenant:profile:update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTenantInput"
      responses:
        "200":
          description: Tenant updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantDetail"
        "403":
          description: 无权限
        "404":
          description: 租户不存在
  /tenant/members:
    get:
      operationId: listTenantMembers
      tags:
        - tenant
      summary: List tenant members
      description: 返回当前租户（基于 X-Tenant-Id）的所有成员列表，需要 tenant:member:read 权限
      responses:
        "200":
          description: Member list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TenantMember"
        "403":
          description: 无权限
  /tenant/members/{membershipId}:
    delete:
      operationId: removeTenantMember
      tags:
        - tenant
      summary: Remove a member from the current tenant
      description: 从当前租户中移除成员（不删除用户账号），需要 tenant:member:remove 权限
      parameters:
        - name: membershipId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Member removed
        "403":
          description: 无权限
        "404":
          description: 成员不存在
  /tenant/members/{membershipId}/roles:
    put:
      operationId: updateMemberRoles
      tags:
        - tenant
      summary: Update member roles
      description: 更新成员的角色列表，需要 tenant:member:update 权限
      parameters:
        - name: membershipId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMemberRolesInput"
      responses:
        "200":
          description: Member updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantMember"
        "403":
          description: 无权限
        "404":
          description: 成员不存在
  /tenant/invitations:
    get:
      operationId: listTenantInvitations
      tags:
        - tenant
      summary: List tenant invitations
      description: 返回当前租户的所有邀请列表，需要 tenant:member:read 权限
      responses:
        "200":
          description: Invitation list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TenantInvitation"
        "403":
          description: 无权限
    post:
      operationId: createTenantInvitation
      tags:
        - tenant
      summary: Invite a user to the current tenant
      description: 通过邮箱邀请新成员加入当前租户，需要 tenant:member:create 权限
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTenantInvitationInput"
      responses:
        "201":
          description: Invitation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantInvitation"
        "400":
          description: 参数错误
        "403":
          description: 无权限
  /tenant/invitations/{invitationId}:
    delete:
      operationId: cancelTenantInvitation
      tags:
        - tenant
      summary: Cancel a tenant invitation
      description: 取消待接受的邀请，需要 tenant:member:create 权限
      parameters:
        - name: invitationId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Invitation cancelled
        "403":
          description: 无权限
        "404":
          description: 邀请不存在
...
  /tenant/roles:
    get:
      operationId: listTenantRoles
      tags:
        - tenant-role
      summary: List tenant roles
      description: 返回当前租户的所有角色列表
      responses:
        "200":
          description: Role list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TenantRole"
        "403":
          description: 无权限

    post:
      operationId: createTenantRole
      tags:
        - tenant-role
      summary: Create a tenant role
      description: 创建新角色，需要 tenant:role:manage 权限
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTenantRoleInput"
      responses:
        "201":
          description: Role created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantRole"
        "400":
          description: 参数错误
        "403":
          description: 无权限
  /tenant/roles/{roleId}:
    get:
      operationId: getTenantRole
      tags:
        - tenant-role
      summary: Get tenant role detail
      description: 返回角色详情
      parameters:
        - name: roleId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Role detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantRole"
        "403":
          description: 无权限
        "404":
          description: 角色不存在

    put:
      operationId: updateTenantRole
      tags:
        - tenant-role
      summary: Update tenant role
      description: 更新角色的显示名称和权限，需要 tenant:role:manage 权限
      parameters:
        - name: roleId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTenantRoleInput"
      responses:
        "200":
          description: Role updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantRole"
        "403":
          description: 无权限
        "404":
          description: 角色不存在

    delete:
      operationId: deleteTenantRole
      tags:
        - tenant-role
      summary: Delete a tenant role
      description: 删除角色（有成员使用时不可删除），需要 tenant:role:manage 权限
      parameters:
        - name: roleId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Role deleted
        "403":
          description: 无权限
        "404":
          description: 角色不存在
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    TenantDetail:
      type: object
      required:
        - tenantId
        - name
        - status
        - roleNames
        - permissionNames
        - createdAt
        - updatedAt
      properties:
        tenantId:
          type: integer
          format: int64
          description: 租户 ID
        name:
          type: string
          description: 租户名称（不可修改）
        displayName:
          type: string
          nullable: true
          description: 租户显示名称
        status:
          type: string
          enum:
            - PENDING_ACTIVATION
            - ACTIVE
            - SUSPENDED
            - DEPROVISIONED
          description: 租户状态
        roleNames:
          type: array
          items:
            type: string
          description: 当前用户在该租户下的角色列表
        permissionNames:
          type: array
          items:
            type: string
          description: 当前用户在该租户下的权限列表
        createdAt:
          type: integer
          format: int64
          description: 创建时间戳（毫秒）
        updatedAt:
          type: integer
          format: int64
          description: 更新时间戳（毫秒）
    UpdateTenantInput:
      type: object
      properties:
        displayName:
          type: string
          nullable: true
          description: 租户显示名称
          maxLength: 200
    TenantMember:
      type: object
      required:
        - id
        - tenantId
        - userId
        - userName
        - status
        - roleNames
      properties:
        id:
          type: integer
          format: int64
          description: 成员资格 ID
        tenantId:
          type: integer
          format: int64
          description: 租户 ID
        userId:
          type: integer
          format: int64
          description: 用户 ID
        userName:
          type: string
          description: 用户名称
        status:
          type: string
          enum:
            - ACTIVE
            - SUSPENDED
            - PENDING_INVITATION
          description: 成员状态
        roleNames:
          type: array
          items:
            type: string
          description: 角色名称列表
    UpdateMemberRolesInput:
      type: object
      required:
        - roleIds
      properties:
        roleIds:
          type: array
          items:
            type: integer
            format: int64
          description: 要分配的角色 ID 列表
    CreateTenantInvitationInput:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
          description: 被邀请人的邮箱地址
        roleIds:
          type: array
          items:
            type: integer
            format: int64
          description: 邀请时指定的角色ID列表
    AcceptTenantInvitationInput:
      type: object
      required:
        - token
      properties:
        token:
          type: string
          description: 邀请 token
    TenantInvitation:
      type: object
      required:
        - id
        - tenantId
        - tenantName
        - email
        - roleIds
        - roleNames
        - status
        - expiresAt
        - createdAt
        - initialAdministrator
      properties:
        id:
          type: integer
          format: int64
          description: 邀请 ID
        tenantId:
          type: integer
          format: int64
          description: 目标租户 ID
        tenantName:
          type: string
          description: 目标租户名称
        email:
          type: string
          description: 被邀请人邮箱
        roleIds:
          type: array
          items:
            type: integer
            format: int64
          description: 邀请时指定的角色 ID 列表
        roleNames:
          type: array
          items:
            type: string
          description: 角色名称列表
        invitedByUserId:
          type: integer
          format: int64
          nullable: true
          description: 发起邀请的用户 ID
        token:
          type: string
          nullable: true
          description: 邀请唯一 token；仅在创建或重发响应中返回一次，列表和接受响应不会返回
        status:
          type: string
          enum:
            - PENDING
            - ACCEPTED
            - CANCELLED
          description: 邀请状态
        expiresAt:
          type: integer
          format: int64
          description: 过期时间戳（毫秒）
        acceptedAt:
          type: integer
          format: int64
          nullable: true
          description: 接受时间戳（毫秒）
        createdAt:
          type: integer
          format: int64
          description: 创建时间戳（毫秒）
        initialAdministrator:
          type: boolean
          description: 是否为待激活租户的首位管理员邀请
    TenantRole:
      type: object
      required:
        - id
        - name
        - label
        - permissionNames
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
          format: int64
          description: 角色 ID
        name:
          type: string
          description: 角色名称（租户内唯一）
        label:
          type: string
          description: 显示名称
        permissionNames:
          type: array
          items:
            type: string
          description: 权限名称列表
        createdAt:
          type: integer
          format: int64
          description: 创建时间戳（毫秒）
        updatedAt:
          type: integer
          format: int64
          description: 更新时间戳（毫秒）
    CreateTenantRoleInput:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: 角色名称（英文/数字/下划线/冒号，1-50字符）
          maxLength: 50
          pattern: "^[a-zA-Z0-9_:]*$"
        label:
          type: string
          nullable: true
          description: 显示名称
          maxLength: 255
        permissionIds:
          type: array
          items:
            type: integer
            format: int64
          description: 要分配的权限 ID 列表
    UpdateTenantRoleInput:
      type: object
      properties:
        label:
          type: string
          nullable: true
          description: 显示名称
          maxLength: 255
        permissionIds:
          type: array
          items:
            type: integer
            format: int64
          description: 要分配的权限 ID 列表
...